test: consolidate operator coverage
