--- tests/test_gate_up_swiglu_kernel.py
@@ -1,28 +1,0 @@
-import pytest
-import torch
-
-from sparsevllm.triton_kernel.gate_up_swiglu import h20_gate_up_swiglu
-
-
-def _is_h20() -> bool:
-    return torch.cuda.is_available() and torch.cuda.get_device_name() == "NVIDIA H20"
-
-
-@pytest.mark.skipif(not _is_h20(), reason="requires NVIDIA H20")
-@pytest.mark.parametrize("intermediate_size", [256, 512])
-def test_h20_gate_up_swiglu_matches_torch(intermediate_size):
-    torch.manual_seed(0)
-    inputs = torch.randn(1, 2048, dtype=torch.bfloat16, device="cuda")
-    weight = 0.02 * torch.randn(
-        2 * intermediate_size,
-        2048,
-        dtype=torch.bfloat16,
-        device="cuda",
-    )
-    projected = torch.nn.functional.linear(inputs, weight)
-    gate, up = projected.chunk(2, dim=-1)
-    expected = torch.nn.functional.silu(gate.float()) * up.float()
-
-    actual = h20_gate_up_swiglu(inputs, weight)
-
-    torch.testing.assert_close(actual.float(), expected, rtol=0.02, atol=0.01)
